IB/hfi1: Make mmap_sem consistent in mmu_rb
The ops->remove() callback was called by hfi1_mmu_unregister() with a
NULL mm argument while holding a spinlock.  In the case of sdma_rb_remove()
this caused it to pass current->mm to hfi1_release_user_pages()

This had 2 problems.  First this would attempt to acquire the mmap_sem
under a spin lock.  Second the use of current->mm is not always guaranteed
to be the proper mm when the fd is being closed.

Rather than depend on this implicit behavior we make the holding of mmap_sem
consistent across all calls to ops->remove in mmu_rb.c.  This then also allows
the correct mm to be used in the remove callback without fear of deadlock.

mmu_rb_remove() in user_exp_recv.c is the only other implementation of
this callback and will need to reference the mm in future patches.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Jim Foraker <foraker1@llnl.gov>