IB/hfi1: Reset QSFP on every run through channel tuning
Through a mistake in the code, the QSFP was reset only every alternate
run through the channel tuning algorithm instead of every time. This was
due to incorrectly resetting the flag that controlled QSFP reset
resulting in using stale status in the channel tuning algorithm. This
patch fixes this issue.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>