IB/hfi1: Consistently call ops->remove outside spinlock
The handler spinlock was only designed to protect the handler rb tree and
there was no need, nor an expectation that a lock will be held while the
users remove is called.  (The user of mmu_rb.c is expected to ensure that
access to their nodes are protected.)

Not only does this clarify the code but a subsequent patch requires that
the remove callback potentially sleep which is now allowed.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>