IB/hfi1: Make use of mm consistent
The hfi1 driver registers a mmu_notifier callback when /dev/hfi1 is opened,
and unregisters it when the device is closed.  The driver assumes that the
close will always happen from the same context as the open, which is
incorrect.  In particular, closes due to SIGKILL or OOM killer activity may
happen from a different context.  In these cases, the wrong mm is passed to
mmu_notifier_unregister(), which causes improper reference counting for the
victim mm, and eventual memory corruption.

Preserve the mm for all open file descriptors and use this mm rather than
current->mm for all memory operations for the lifetime of that fd.

Furthermore, if registration fails there is no reason to keep the handler
object around.  So free the handler object rather than add it to he list
to prevent any mmu_notifier operations, including unregister, when
registration fails.

Suggested-by: Jim Foraker <foraker1@llnl.gov>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>