IB/hfi1: set_intr_bits uses incorrect source for register modification
HFI IRQ enable bits are not being set correctly.  Send context error
and DC IRQs are not being enabled correctly.  In addition, send
context error IRQs are not being delivered.

Because of this, send context errors are not being handled correctly
when they occur.

When setting the IRQ bits, if an IRQ range is used, and the last bit
is on a register boundary (bit 63), the calculated index for the
final register modification is incorrect (index + 1 vs. index).

The incorrect index calculation causes incorrect IRQ bits to be set.
In this case the send context error IRQ is NOT enabled.

Fix by using the 'last' value rather than the counted 'src' value to
determine the final index to use.  This satisfies all cases.

This commit is found in the IFS release 10.9.1

Change-Id: I4c73aeee87e8e7a4023f15992951eae1a52f9fb1
Fixes: a2f7bbdc2dba ("IB/hfi1: Rework the IRQ API to be more flexible")
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>