IB/hfi1: Disable HW IRQ before freeing PCI IRQ
The pci_free_irq() requires that the HW IRQ is disabled before
it can be called.

This is not being done for netdev contexts.

Zero (0) is a valid msix_intr value. Initialize the msix_intr to an
invalid value so the free path can be checked easily.

Correct the call order to meet the pci_free_irq() requirement.

This commit is found in the IFS release 10.9.2

Fixes: a1a482711a53 ("IB/hfi1: Introduce OPA accelerated IPoIB netdev RX")
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Sadanand Warrier <sadanand.warrier@intel.com>
Signed-off-by: Michael J. Ruhl <michael.j.ruhl@intel.com>