IB/hfi1: Restore affinity for non-ipoib_accel
The 10.9 release always reserves hints for netdev
contexts even when acceleration is turned off.  This
flaw reduces the number of cores to which sdma interrupts
are hinted, impacting performance on low core count cpus.

Adjust the affinity initialization to not do the
reservation when acceleration is disabled.

Change-Id: Ia0c6dabc20fd45c95b8aa3932b4e3d671e356ce7
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>

Conflicts:
	drivers/infiniband/hw/hfi1/ipoib_main.c

Resolved by John Fleck

This commit is found in the IFS release 10.9.1

Signed-off-by: John Fleck <john.fleck@intel.com>