IB/hfi1: Share processor affinity hints between SDMA and Netdev IRQs
Netdev IRQ hints are added to the recv_intr mask when acceleration is
enabled and are mutually exclusive from SDMA in the def_intr mask.
This causes there to be a reduced number of available cores for SDMA thus
reducing the number of cores to which SDMA interrupts are hinted.
This impacts performance on low core count cpus.

Sharing the affinity hints between SDMA and Netdev contexts solves this
problem.

Netdev IRQs use the recv_intr mask for affinity hints.
Recvctxt IRQs need to be on an isolated CPU for best performance.
Netdev IRQs do not have this requirement.
Netdev IRQs now use the def_intr mask to share the CPU requirements
with SDMA.

A new msix type of IRQ_NETDEVCTXT has been introduced to identify
Netdev IRQ requests which use the def_intr mask for affinity hint
assignments.

This commit is found in the IFS release 10.9.2

Fixes: a1a482711a53 ("IB/hfi1: Introduce OPA accelerated IPoIB netdev RX")
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Gary Leshner <Gary.S.Leshner@intel.com>