IB/hfi1: Add limit test for RC/UC send via loopback
Fix potential memory corruption and panic in loopback for
IB_WR_SEND varaiants.

The code blindly assumes the posted length will fit in the
fetched rwqe, which is not a valid assumption.

Fix by adding a limit test, and triggering the appropriate
send completion and putting the QP in an error state.  This
mimics the handling for non-loopback QPs.

[ported to older kernels where the loopback was in each driver]

This commit is found in the IFS release 10.9.2

Cc: <stable@vger.kernel.org> #v4.9+
Fixes: 7724105686e7 ("IB/hfi1: add driver files")
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>