IB/hfi1: Close hole for enable/disable of AIP contexts
AIP contexts are born enabled and with interrupts enabled
but before the napi rx is enabled to process the packets.

This creates a window of time where ipoib UD packets from
senders with cached arp information can send packets that will land
in a receive context but not be processed for a significant amount of time.
In the extreme case, backpressure from a full send context can cause packets
to fill the RX_BUFFER.  This can run the neighbor switch out of credits,
causing a deadlock that impacts all contexts.

Fix by insuring the AIP context's receive control register and
the receive interrupts are not enabled until the napi rx is
ready to process the packets. Conversely, on disable, insure the context
is disabled along with interrupts.

This commit is found in the IFS release 10.9.1

Fixes: 6d27f9cd3048 ("IB/hfi1: Introduce OPA accelerated IPoIB netdev RX")
Reviewed-by: Gary Leshner <Gary.S.Leshner@intel.com>
Reviewed-by: Sadanand Warrier <sadanand.warrier@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>