IB/hfi1: Ensure correct mm is used at all times
Two earlier bug fixes have created a security problem in the hfi1
driver. One fix aimed to solve an issue where current->mm was not valid
when closing the hfi1 cdev. It attempted to do this by saving a cached
value of the current->mm pointer at file open time. This is a problem if
another process with access to the FD calls in via write() or ioctl() to
pin pages via the hfi driver. The other fix tried to solve a use after
free by taking a reference on the mm.

To fix this correctly we use the existing cached value of the mm in the
mmu notifier. Now we can check in the insert, evict, etc. routines that
current->mm matched what the notifier was registered for. If not, then
don't allow access. The register of the mmu notifier will save the mm
pointer.

Since in do_exit() the exit_mm() is called before exit_files(), which
would call our close routine a reference is needed on the mm. We rely on
the mmgrab done by the registration of the notifier, whereas before it was
explicit. The mmu notifier deregistration happens when the user context is
torn down, the creation of which triggered the registration.

Also of note is we do not do any explicit work to protect the interval
tree notifier. It doesn't seem that this is going to be needed since we
aren't actually doing anything with current->mm. The interval tree
notifier stuff still has a FIXME noted from a previous commit that will be
addressed in a follow on patch.

Backport notes:

This patch requires tracking of two situations:
- If the underlying mmu_notifier_register() does not do an mmgrab()
  * NO_NOTIFIER_REG_GRAB
- If the struct struct mmu_notifier does not have an mm pointer
  * NO_MMU_NOTIFIER_MM

These are independent flags and need to be specified for each distro.

Backport adjustments are made to do an mmgrab() in mmu_rb.c and
specifying the source for the mm pointer for situations where a callback is on on
the "current" process that did the mmu_notifier_register().

To support GPU builds the following changes are needed:
- hfi1_mmu_rb_register_gpu() is added to support GPU rb registration
  * The routine is a copy paste of the hfi1_mmu_rb_register() without
    the notifier registration
  * These calls do an implict mmgrab() and save current properly
- A new member in mmu_rb_handler, reserved, tracks if the notifer is
registered
  * registered is true for non-GPU and false for GPU
- hfi1_mmu_rb_unregister() is changed to use registered if the notifier
  registration has occurred and unregister and do the mm reference stuff
  * The mmdrop() is called for GPU to pair with the mmgrab() in the GPU
    registration
- file_ops.c no longer needs linux/sched/mm.h since that is now only needed
  in mmu_rb.c with the appropriate guard on the include

Cc: <stable@vger.kernel.org>
Fixes: e0cf75deab81 ("IB/hfi1: Fix mm_struct use after free")
Fixes: 3faa3d9a308e ("IB/hfi1: Make use of mm consistent")
Link: https://lore.kernel.org/r/20201125210112.104301.51331.stgit@awfm-01.aw.intel.com
Suggested-by: Jann Horn <jannh@google.com>
Reported-by: Jason Gunthorpe <jgg@nvidia.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>