Reconcile with 10.11.0.1.2 released code
This code reflects 10.11.0.1.2, NOT 10.11.0.2.2. That was
most likely a typo in the original commit.

The SHA the code comes from is:
a596bd8c81a2773c13519fcff14abbb9ad812f01

Also there are some leftover work in progress files and build scripts that do not
belong in this published repo. Remove that stuff.

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>